+++ examples/datasets/plot_ice_margin.py
@@ -24,6 +24,9 @@ with hyoga.open.example('pism.alps.out.2d.nc') as ds:
     # add coastline and rivers
     ax = ds.hyoga.plot.natural_earth()
 
+    # add scale bar
+    ds.hyoga.plot.scale_bar()
+
 # set title
 ax.set_title('Ice margin')