+++ doc/whatsnew.rst
@@ -21,6 +21,17 @@
 What's new
 ==========
 
+.. _v0.2.2:
+
+v0.2.2 (unreleased)
+-------------------
+
+New features
+------------
+
+- Add plot method :meth:`.Dataset.hyoga.plot.scale_bar` and a new example for
+  automatically sized, anchored scale bars (:issue:`16`, :pull:`44`).
+
 .. _v0.2.1:
 
 v0.2.1 (1 Dec. 2022)